+++ packages/coin-support-sia/src/operations/broadcastTransaction/types.ts
@@ -1,0 +1,8 @@
+import { IBroadcastTransactionParams } from '@cypherock/coin-support-interfaces';
+
+import { IPreparedSiaTransaction } from '../transaction';
+
+export interface IBroadcastSiaTransactionParams
+  extends IBroadcastTransactionParams<string> {
+  transaction: IPreparedSiaTransaction;
+}