+++ packages/coins/src/aggregate.ts
@@ -6,6 +6,7 @@ import { nearCoinList, NearId } from './near';
 import { solanaCoinList, SolanaId } from './solana';
 import { starknetCoinList, StarknetId } from './starknet';
 import { stellarCoinList, StellarId } from './stellar';
+import { siaCoinList, SiaId } from './sia';
 import { tronCoinList, ITronTrc20Token } from './tron';
 import { ICoinInfo } from './types';
 import { xrpCoinList, XrpId } from './xrp';
@@ -19,6 +20,7 @@ export const coinList: Record<
   | StarknetId
   | IcpId
   | StellarId
+  | SiaId
   | CantonId,
   ICoinInfo
 > = {
@@ -31,6 +33,7 @@ export const coinList: Record<
   ...starknetCoinList,
   ...icpCoinList,
   ...stellarCoinList,
+  ...siaCoinList,
   ...cantonCoinList,
 };
 
@@ -43,6 +46,7 @@ export type CoinTypes =
   | StarknetId
   | IcpId
   | StellarId
+  | SiaId
   | CantonId;
 export type TokenTypes =
   | IEvmErc20Token