+++ packages/cysync-core/src/components/CoinIcon.tsx
@@ -6,6 +6,7 @@ import {
   NearIdMap,
   XrpIdMap,
   StellarIdMap,
+  SiaIdMap,
   CantonIdMap,
 } from '@cypherock/coins';
 import {
@@ -30,6 +31,7 @@ import {
   Image,
   UtilsProps,
   StellarIcon,
+  SiacoinIcon,
   CantonIcon,
   CbtcIcon,
 } from '@cypherock/cysync-ui';
@@ -70,6 +72,7 @@ const coinToIconMap: Record<string, React.FC<IconProps> | undefined> = {
   [SolanaIdMap.solana]: SolanaIcon,
   [XrpIdMap.xrp]: XrpIcon,
   [StellarIdMap.stellar]: StellarIcon,
+  [SiaIdMap.sia]: SiacoinIcon,
   [CantonIdMap.canton]: CantonIcon,
 } as Record<string, React.FC<IconProps> | undefined>;