+++ packages/cysync-core/src/dialogs/Send/Dialogs/Components/AddressAndAmountSection/index.tsx
@@ -136,6 +136,17 @@ const StellarAddressAndAmount: React.FC<AnaProps> = ({
   </Container>
 );
 
+const SiaAddressAndAmount: React.FC<AnaProps> = ({
+  disableInputs,
+  providerName,
+}) => (
+  <Container px={5} py="12px">
+    <SingleTransaction
+      disableInputs={disableInputs}
+      providerName={providerName}
+    />
+  </Container>
+);
 const CantonAddressAndAmount: React.FC<AnaProps> = ({
   disableInputs,
   providerName,
@@ -161,6 +172,7 @@ XrpAddressAndAmount.defaultProps = defaultAnaProps;
 StarknetAddressAndAmount.defaultProps = defaultAnaProps;
 IcpAddressAndAmount.defaultProps = defaultAnaProps;
 StellarAddressAndAmount.defaultProps = defaultAnaProps;
+SiaAddressAndAmount.defaultProps = defaultAnaProps;
 CantonAddressAndAmount.defaultProps = defaultAnaProps;
 
 const anaInputMap: Record<CoinFamily, React.FC<any>> = {
@@ -173,6 +185,7 @@ const anaInputMap: Record<CoinFamily, React.FC<any>> = {
   starknet: StarknetAddressAndAmount,
   icp: IcpAddressAndAmount,
   stellar: StellarAddressAndAmount,
+  sia: SiaAddressAndAmount,
   canton: CantonAddressAndAmount,
 };