+++ packages/cysync-core/src/dialogs/Send/hooks/useLabelSuffix.ts
@@ -21,6 +21,7 @@ export const useLabelSuffix = () => {
     starknet: getDefaultSuffix,
     icp: getDefaultSuffix,
     stellar: getDefaultSuffix,
+    sia: getDefaultSuffix,
     canton: getDefaultSuffix,
   };
   const getFeeLabelSuffix = (selectedAccount?: IAccount) => {