+++ apps/desktop/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cypherock/cysync-desktop
 
+## 2.3.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-desktop-ui@0.2.6
+
+## 2.3.6-pre.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @cypherock/cysync-desktop-ui@0.2.6-pre.0
+
 ## 2.3.5
 
 ### Patch Changes